+++ CHANGELOG.md
@@ -10,6 +10,24 @@
 
 ### Fixed
 
+## 3.4.0 - 2024-01-21
+
+### Added
+
+- Settings option to enable or disable the grouping of tools
+- Settings option to sort the tools menu alphabetically
+
+### Changed
+
+- By default, the tools menu is a flat alphabetical list. The old behavior (e.g., grouping of nodes) can be restored through the settings.
+- Removed the setting to hide the tool window menu on a tool selection. The selection mechanism now distinguishes between an automatic search result selection (the menu remains visible) and an user selection (a menu is hidden).
+- Renamed tool "Java Text Escape/Unescape" to "Java String Escape/Unescape"
+- Renamed tool "Code Formatting Converter" to "Code Format Converter"
+
+### Fixed
+
+- Tool "Code Format Converter" wasn't working correctly
+
 ## 3.3.0 - 2024-01-15
 
 ### Added